♻️(back) rename video task to s3 task
In order to be more explicit about what the task is doing, rename the
video task to s3 task. It's also avoid circular import with video
tasks.
